refactor: combine Origin and Redirect
