+++ src/Appwrite/Network/Client.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Appwrite\Network;
+
+class Client
+{
+    public const TYPE_UNKNOWN = 'unknown';
+    public const TYPE_WEB = 'web';
+    public const TYPE_FLUTTER_IOS = 'flutter-ios';
+    public const TYPE_FLUTTER_ANDROID = 'flutter-android';
+    public const TYPE_FLUTTER_MACOS = 'flutter-macos';
+    public const TYPE_FLUTTER_WINDOWS = 'flutter-windows';
+    public const TYPE_FLUTTER_LINUX = 'flutter-linux';
+    public const TYPE_FLUTTER_WEB = 'flutter-web';
+    public const TYPE_APPLE_IOS = 'apple-ios';
+    public const TYPE_APPLE_MACOS = 'apple-macos';
+    public const TYPE_APPLE_WATCHOS = 'apple-watchos';
+    public const TYPE_APPLE_TVOS = 'apple-tvos';
+    public const TYPE_ANDROID = 'android';
+    public const TYPE_UNITY = 'unity';
+    public const TYPE_REACT_NATIVE_IOS = 'react-native-ios';
+    public const TYPE_REACT_NATIVE_ANDROID = 'react-native-android';
+}
